SilenceBase in lib-builtin-effects
