+++ src/effects/Silence.cpp
@@ -17,48 +17,9 @@
 #include "LoadEffects.h"
 
 #include "ShuttleGui.h"
-#include "WaveTrack.h"
 #include "../widgets/NumericTextCtrl.h"
 
-const ComponentInterfaceSymbol SilenceBase::Symbol
-/* i18n-hint: noun */
-{ XC("Silence", "generator") };
-
 namespace{ BuiltinEffectsModule::Registration< EffectSilence > reg; }
-
-SilenceBase::SilenceBase()
-{
-   SetLinearEffectFlag(true);
-}
-
-SilenceBase::~SilenceBase()
-{
-}
-
-// ComponentInterface implementation
-
-ComponentInterfaceSymbol SilenceBase::GetSymbol() const
-{
-   return Symbol;
-}
-
-TranslatableString SilenceBase::GetDescription() const
-{
-   return XO("Creates audio of zero amplitude");
-}
-
-ManualPageID SilenceBase::ManualPage() const
-{
-   return L"Silence";
-}
-
-
-// EffectDefinitionInterface implementation
-
-EffectType SilenceBase::GetType() const
-{
-   return EffectTypeGenerate;
-}
 
 // Effect implementation
 
@@ -102,12 +63,5 @@ bool EffectSilence::TransferDataFromWindow(EffectSettings &settings)
 {
    settings.extra.SetDuration(mDurationT->GetValue());
 
-   return true;
-}
-
-bool SilenceBase::GenerateTrack(
-   const EffectSettings &settings, WaveTrack &tmp)
-{
-   tmp.InsertSilence(0.0, settings.extra.GetDuration());
    return true;
 }
