+++ src/effects/Silence.h
@@ -13,34 +13,10 @@
 #ifndef __AUDACITY_EFFECT_SILENCE__
 #define __AUDACITY_EFFECT_SILENCE__
 
-#include "Generator.h"
+#include "SilenceBase.h"
 #include "StatefulEffectUIServices.h"
 
 class NumericTextCtrl;
-
-class SilenceBase : public Generator
-{
-public:
-   static const ComponentInterfaceSymbol Symbol;
-
-   SilenceBase();
-   virtual ~SilenceBase();
-
-   // ComponentInterface implementation
-
-   ComponentInterfaceSymbol GetSymbol() const override;
-   TranslatableString GetDescription() const override;
-   ManualPageID ManualPage() const override;
-
-   // EffectDefinitionInterface implementation
-
-   EffectType GetType() const override;
-
-protected:
-   // Generator implementation
-
-   bool GenerateTrack(const EffectSettings &settings, WaveTrack &tmp) override;
-};
 
 class EffectSilence final : public SilenceBase, public StatefulEffectUIServices
 {
